Add files via upload
First, we looked for missing values, dropped the duplicate values and performed preliminary EDA.
We check for multicollinearity with the help of VIF and split the data into training and testing sets, we applied Linear Regression.
We fit the Linear Regression Model in our dataset and find the  values, RMSE for the training set that gives approximately 89.6% of the variance in the dependent variable. So we conclude that the model provides a reasonably good fit to the data.